requirements.txt
biopython
matplotlib
numpy
pandas
biopython
matplotlib
nupack
kubernetes
botocore
boto3
